refactor : cards controllers now use class
